Add Maven Project Structure
The project structure for a Spring Boot application maintaining a Person's REST API is established. Various files and directories relevant to Spring Boot and Maven are added, including pom.xml, Spring Boot Application (Application.java), tests, Docker Compose, properties, application servers, and more. The application aims to practice Hexagonal Architecture concepts with Java and Spring. Lombok has been used for reducing boilerplate code in the application and the Maven Wrapper allows the project to be built with the correct version of Maven without it being installed on the target platform.