Refactor persistence layer and test code
The AddressEntity, PersonEntity, and test classes have been refactored. The imported validation annotations (@NotNull) were removed and replaced with Hibernate-specific annotations in AddressEntity and PersonEntity. In addition, the CascadeType of the @ManyToOne annotation in PersonEntity was set to PERSIST. The test class switched from @SpringBootTest to @DataJpaTest, and new assertions were added to validate dates.